chore: Renamed file to match class' name
