+++ tests/configuration/io/test_koswat_surroundings_importer.py
@@ -7,7 +7,7 @@ from koswat.configuration.io.ini.koswat_general_ini_fom import (
     InfrastructureSectionFom,
     SurroundingsSectionFom,
 )
-from koswat.configuration.io.koswat_surroundings_importer import (
+from koswat.configuration.io.surroundings_wrapper_collection_importer import (
     SurroundingsWrapperCollectionImporter,
 )
 from koswat.configuration.settings.koswat_general_settings import (
